webpack ganache-cli for npm distribution
