+++ .npmignore
@@ -2,10 +2,15 @@
 .tern-project
 .tern-port
 .vscode
-.travis.yml
 .eslintrc.js
 *.log
-TODO
-docker/build
+build/ganache-core.docker.cli.js
+build/ganache-core.docker.cli.js.map
+webpack
 perf
-npm-shrinkwrap.json.bak
+.dockerignore
+Dockerfile
+ISSUE_TEMPLATE.md
+cli.js
+lib.js
+args.js
