+++ package.json
@@ -1,38 +1,33 @@
 {
   "name": "ganache-cli",
-  "version": "6.2.2-beta.8",
+  "version": "6.2.2-beta.9",
   "bin": {
-    "ganache-cli": "cli.js"
+    "ganache-cli": "./build/ganache-core.node.cli.js"
   },
-  "main": "lib.js",
-  "directories": {
-    "lib": "./lib"
-  },
+  "main": "./build/ganache-core.node.cli.js",
   "scripts": {
     "start": "node cli.js",
-    "build": "webpack-cli --config ./docker/webpack.config.js",
-    "prepublishOnly": "npm shrinkwrap && mv npm-shrinkwrap.json npm-shrinkwrap.json.bak && npm prune --only=prod && npm install json --no-save -E && npm shrinkwrap",
-    "postshrinkwrap": "json -f npm-shrinkwrap.json -I -e \"d=this.dependencies;g=d['ganache-core'];h=g.dependencies;h.keccakjs && (delete h.keccakjs.requires.sha3);delete g.requires.sha3;delete d.sha3;delete h.sha3;h['scrypt.js'] && (delete h['scrypt.js'].requires.scrypt);delete g.requires.scrypt;delete d.scrypt;\"",
-    "postpublish": "rm npm-shrinkwrap.json && mv npm-shrinkwrap.json.bak npm-shrinkwrap.json && npm install"
+    "build": "webpack-cli --config ./webpack/webpack.node.config.js",
+    "docker": "docker build -t trufflesuite/ganache-cli . && docker run -p 8545:8545 trufflesuite/ganache-cli",
+    "prepublishOnly": "npm run build"
   },
   "dependencies": {
-    "bn.js": "4.11.8",
-    "ganache-core": "2.3.2-beta.2",
     "source-map-support": "0.5.9",
     "yargs": "11.1.0"
   },
+  "bundleDependencies": [
+    "source-map-support",
+    "yargs"
+  ],
   "devDependencies": {
-    "async": "^2.6.1",
-    "json": "9.0.6",
+    "bn.js": "4.11.8",
+    "ganache-core": "2.3.2-beta.2",
     "node-loader": "^0.6.0",
     "on-build-webpack": "^0.1.0",
     "prepend-file": "^1.3.1",
     "shebang-loader": "0.0.1",
-    "web3": "^1.0.0-beta.36",
     "webpack": "^4.18.0",
-    "webpack-cli": "3.1.0"
-  },
-  "optionalDependencies": {
+    "webpack-cli": "3.1.0",
     "scrypt": "6.0.3",
     "sha3": "1.2.2"
   },
